+++ src/org/uecide/gui/swing/SketchTree.java
@@ -4,11 +4,14 @@ import org.uecide.Context;
 import org.uecide.Sketch;
 
 import javax.swing.JPopupMenu;
+import javax.swing.JMenuItem;
 import javax.swing.JTree;
 import javax.swing.tree.TreePath;
 
 import java.awt.event.MouseListener;
 import java.awt.event.MouseEvent;
+import java.awt.event.ActionListener;
+import java.awt.event.ActionEvent;
 
 public class SketchTree extends JTree implements MouseListener {
     Context ctx;
@@ -42,6 +45,15 @@ public class SketchTree extends JTree implements MouseListener {
 
             SketchTreeNodeBase node = (SketchTreeNodeBase)selPath.getLastPathComponent();
             JPopupMenu menu = node.getPopupMenu();
+            JMenuItem refresh = new JMenuItem("Refresh");
+            refresh.addActionListener(new ActionListener() {
+                public void actionPerformed(ActionEvent evt) {
+                    if (node.updateChildren()) {
+                        model.reload(node);
+                    }
+                }
+            });
+            menu.add(refresh);
             if (menu != null) {
                 menu.show(this, evt.getX(), evt.getY());
             }