+++ src/org/uecide/gui/swing/SketchTreeOutputNode.java
@@ -1,6 +1,8 @@
 package org.uecide.gui.swing;
 
 import org.uecide.Context;
+import org.uecide.ContextEvent;
+import org.uecide.ContextEventListener;
 import org.uecide.FileType;
 import org.uecide.SketchFile;
 
@@ -10,27 +12,60 @@ import javax.swing.JPopupMenu;
 import javax.swing.JLabel;
 import javax.swing.JTree;
 
+import java.io.File;
 import java.io.IOException;
 
 import java.awt.Font;
 import java.awt.Component;
 
-public class SketchTreeOutputNode extends SketchTreeNodeBase {
+import java.util.Enumeration;
+
+public class SketchTreeOutputNode extends SketchTreeNodeBase implements ContextEventListener {
     public SketchTreeOutputNode(Context c, SketchTreeModel m) {
         super(c, m, "Output");
         updateChildren();
+        ctx.listenForEvent("buildFileUpdated", this);
+        ctx.listenForEvent("buildFileRemoved", this);
+        ctx.listenForEvent("buildFileAdded", this);
     }
 
     public boolean updateChildren() {
-        removeAllChildren();
-        for (SketchFile f : ctx.getSketch().getSketchFiles().values()) {
-            if (f.getGroup() == FileType.GROUP_HEADER) {
-                SketchSourceFileNode sfn = new SketchSourceFileNode(ctx, model, f);
-                add(sfn);
+        File file = ctx.getBuildDir();
+
+        boolean somethingRemoved = false;
+        boolean hasBeenModified = false;
+
+        if (file.isDirectory()) {
+            File[] files = file.listFiles();
+
+            do {
+                somethingRemoved = false;
+                for (Enumeration e = children(); e.hasMoreElements();) {
+                    GenericFileNode child = (GenericFileNode)e.nextElement();
+                    if (!inFileArray(child.getFile(), files)) {
+                        remove(child);
+                        somethingRemoved = true;
+                        hasBeenModified = true;
+                        break;
+                    }
+                }
+            } while (somethingRemoved);
+
+            for (File f : files) {
+                if (!hasChildFile(f)) {
+                    GenericFileNode sfn = new GenericFileNode(ctx, model, f);
+                    add(sfn);
+                    hasBeenModified = true;
+                }
+            }
+            for (Enumeration e = children(); e.hasMoreElements();) {
+                GenericFileNode child = (GenericFileNode)e.nextElement();
+                if (child.updateChildren()) {
+                    hasBeenModified = true;
+                }
             }
         }
-        model.reload(this);
-        return true;
+        return hasBeenModified;
     }
 
     public ImageIcon getIcon(JTree tree) throws IOException {
@@ -55,6 +90,28 @@ public class SketchTreeOutputNode extends SketchTreeNodeBase {
         }
 
         return original;
+    }
+
+    public void contextEventTriggered(ContextEvent evt) {
+        if (updateChildren()) {
+            model.reload(this);
+        }
+    }
+
+
+    boolean inFileArray(File needle, File[] haystack) {
+        for (File f : haystack) {
+            if (f == needle) return true;
+        }
+        return false;
+    }
+
+    boolean hasChildFile(File f) {
+        for (Enumeration e = children(); e.hasMoreElements();) {
+            GenericFileNode child = (GenericFileNode)e.nextElement();
+            if (child.getFile() == f) return true;
+        }
+        return false;
     }